Implement calculator functionality and refactor code
In this commit, methods for a simulated calculator are introduced. Click handlers are added to HTML elements in home.page.html to perform actions such as addition and sign toggling. The home.page.ts methods handle the logic behind these actions. Code is also refactored to include detailed comments and more streamlined construction.